reduce tests size + fix netcore builds
